requirements.txt
Django==4.0.5,<4.1
djangorestframework==3.13.1,<3.14
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.15.1,<0.16
Pillow>=8.2.0,<8.3.0
celery==5.2.6
redis==4.2.2